Make crypto and compression JVM-only
